feat(cloudwatchlogs) Ensure that log groups passed in are created
When provisioning lambdas the cloudwatch log group assigned to that function is not created till the first invoke, this makes subscribing hard as the group names passed in may not exist yet.

This PR adds a new custom cloudformation resource which attempts to create the log group, which works around this issue.